ibdb/requirements.txt
flask
flask-sqlalchemy
flask-graphql
furl
psycopg2-binary
graphene_sqlalchemy>2.0
